+++ src/models/resource.js
@@ -14,7 +14,7 @@ export default types.model('Resource', {
   tags: types.optional(types.array(types.string), []),
   images: types.optional(types.array(types.string), []),
   status: types.optional(types.enumeration(['loading', 'finished']), 'finished'),
-  type: types.optional(types.string, ''),
+  types: types.array(types.string),
   series: types.optional(types.string, ''),
   createdAt: types.optional(types.string, ''),
   words: types.optional(types.number, 0),